pyproject.toml
[tool.poetry]
name = "hw-drf"
version = "0.1.0"
description = ""
authors = ["Serg"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
